requirements.txt
discord.py
requests
aiosqlite
aiohttp
psutil

